// types/user.ts
export interface User {
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber: string;
  addressLine1: string;
  addressLine2: string;
  addressCity: string;
  addressState: string;
  addressPostalCode: string;
  addressCountryCode: string;
  birthDate: string;
  income: string;
  documentSsn: string;
}